+++ test/lib/elements/sleep.mli
@@ -1,4 +1,0 @@
-(** An implementation of sleep for Picos *)
-
-val sleepf : float -> unit
-(** *)