+++ lib/picos_domain/picos_domain.ocaml4.ml
@@ -1,5 +1,6 @@
 let at_exit = Stdlib.at_exit
 let recommended_domain_count () = 1
+let is_main_domain () = true
 
 module DLS = struct
   type 'a state = Value of { mutable value : 'a } | Default of (unit -> 'a)